renamed reload command file
